+++ src/promptflow/promptflow/__init__.py
@@ -7,7 +7,6 @@ from promptflow._core.metric_logger import log_metric
 
 # flake8: noqa
 from promptflow._core.tool import ToolProvider, tool
-from promptflow._core.tracer import trace
 
 # control plane sdk functions
 from promptflow._sdk._load_functions import load_flow, load_run
@@ -20,4 +19,4 @@ log_flow_metric = log_metric
 
 __version__ = VERSION
 
-__all__ = ["PFClient", "load_flow", "load_run", "log_metric", "ToolProvider", "tool", "trace"]
+__all__ = ["PFClient", "load_flow", "load_run", "log_metric", "ToolProvider", "tool"]
