+++ src/promptflow/promptflow/_core/openai_injector.py
@@ -13,9 +13,9 @@ from importlib.metadata import version
 import openai
 
 from promptflow._core.operation_context import OperationContext
-from promptflow.contracts.trace import TraceType
+from promptflow.tracing._trace import _traced_async, _traced_sync
+from promptflow.tracing.contracts.trace import TraceType
 
-from .tracer import _traced_async, _traced_sync
 
 USER_AGENT_HEADER = "x-ms-useragent"
 PROMPTFLOW_HEADER = "ms-azure-ai-promptflow"
