+++ src/promptflow/promptflow/tracing/contracts/__init__.py
@@ -3,8 +3,3 @@
 # ---------------------------------------------------------
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
-
-from ._start_trace import start_trace
-from ._trace import trace
-
-__all__ = ["start_trace", "trace"]
