+++ src/promptflow/promptflow/tracing/contracts/trace.py
