+++ src/promptflow/tests/executor/e2etests/test_executor_execution_failures.py
@@ -23,6 +23,8 @@ Exception: In raise_exception: dummy_input
 The above exception was the direct cause of the following exception:
 
 Traceback (most recent call last):
+in wrapped
+    output = func(*args, **kwargs)
 sync_fail.py", line 13, in raise_an_exception
     raise Exception(f"In tool raise_an_exception: {s}") from e
 Exception: In tool raise_an_exception: dummy_input
@@ -39,6 +41,8 @@ Exception: In raise_exception_async: dummy_input
 The above exception was the direct cause of the following exception:
 
 Traceback (most recent call last):
+in wrapped
+    output = await func(*args, **kwargs)
 in raise_an_exception_async
     raise Exception(f"In tool raise_an_exception_async: {s}") from e
 Exception: In tool raise_an_exception_async: dummy_input
