+++ src/promptflow/tests/test_configs/eager_flows/dummy_flow_with_trace/flow_with_trace.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from promptflow import trace
+from promptflow.tracing import trace
 
 
 @trace
