+++ src/promptflow/tests/test_configs/flows/flow_with_trace/greetings.py
@@ -1,5 +1,6 @@
 from time import sleep
-from promptflow import tool, trace
+from promptflow import tool
+from promptflow.tracing import trace
 
 
 @trace
