+++ src/promptflow/tests/test_configs/flows/flow_with_trace_async/greetings.py
@@ -1,6 +1,6 @@
 import asyncio
-from time import sleep
-from promptflow import tool, trace
+from promptflow import tool
+from promptflow.tracing import trace
 
 
 @trace
